abstract runner logic out + create stubs for new DB's (#1011)
