+++ tests/worker_coordinator/runner_test.py
@@ -33,7 +33,7 @@ import opensearchpy
 import pytest
 from osbenchmark.database.clients.opensearch import opensearch as client
 from osbenchmark import exceptions
-from osbenchmark.worker_coordinator import runner
+from osbenchmark.worker_coordinator import runners as runner
 from tests import run_async, as_future
